+++ src/mock/point-mock.js
@@ -1,6 +1,6 @@
 import dayjs from 'dayjs';
-import {getRandomInteger, generateId} from '../utils/common.js';
-import {POINT_TYPES, MAX_DAYS_GAP, MAX_MINUTES_GAP, POINT_CITIES, OFFER_TITLES_TYPES, TEXT_DESCRIPTION, URL_PHOTO, MAX_COUNT_SENTENCES, MAX_COUNT_PHOTOS, MAX_PRICE_OFFER} from '../const.js';
+import {getRandomInteger, generateId, generateOffers, generateDestination} from '../utils/common.js';
+import {POINT_TYPES, MAX_DAYS_GAP, MAX_MINUTES_GAP, POINT_CITIES} from '../const.js';
 
 const generateTypePoint = () => {
   const randomIndex = getRandomInteger(0, POINT_TYPES.length - 1);
@@ -28,75 +28,6 @@ const generateDate = () => {
 const generateCityPoint = () => {
   const randomIndex = getRandomInteger(0, POINT_CITIES.length - 1);
   return POINT_CITIES[randomIndex];
-};
-
-const generateTypePointOffers = (typePoint) => {
-  const offersTitlesTypes = OFFER_TITLES_TYPES.filter((item) => item.point === typePoint);
-  const offers = offersTitlesTypes.map((item) => ({
-    title: item.title,
-    type: item.type,
-    price: String(getRandomInteger(1, MAX_PRICE_OFFER)),
-    checked: Boolean(getRandomInteger(0, 1)),
-  }));
-
-  return offers;
-};
-
-const generateOffers = (typePoint) => ({
-  type: typePoint,
-  offers: generateTypePointOffers(typePoint),
-});
-
-const generateDescription = (countSentences = MAX_COUNT_SENTENCES) => {
-  const sentences = TEXT_DESCRIPTION.split('. ');
-
-  const randomCount = getRandomInteger(1, countSentences);
-  const sentencesSet = new Set();
-
-  while (sentencesSet.size < randomCount){
-    const randomIndex = getRandomInteger(0, sentences.length - 1);
-    if (!sentencesSet.has(sentences[randomIndex])) {
-      sentencesSet.add(sentences[randomIndex]);
-    }
-  }
-
-  let description = '';
-  for (const value of sentencesSet) {
-    description += `${value}. `;
-  }
-
-  return description;
-};
-
-const checkPicture = (picture, pictures) => pictures.find((item) => item.src === picture.src);
-
-const generatePictures = () => {
-
-  const randomCount = getRandomInteger(0, MAX_COUNT_PHOTOS - 1);
-  const pictures = [];
-
-  while (pictures.length < randomCount) {
-    const randomIndex = getRandomInteger(0, MAX_COUNT_PHOTOS - 1);
-    const picture = {
-      src: `${URL_PHOTO}${randomIndex}`,
-      description: generateDescription(1),
-    };
-    if (!checkPicture(picture, pictures)) {
-      pictures.push(picture);
-    }
-  }
-
-  return pictures;
-};
-
-const generateDestination = (city) => {
-  const description = generateDescription();
-  const pictures = generatePictures();
-  return {
-    description,
-    city,
-    pictures,
-  };
 };
 
 const getDuration = (startDateTime, endDateTime) => {
